Revert "Merge pull request #903 from cisco/feature/mctversion"
This reverts commit d24e75b852bc5b6fa3aef797e9187762494ce083, reversing
changes made to 2d4e4625a3654cfc276ad483e5e75bf002864030.

The previous PR made various adjustments to the code based on wrong assumptions -
1.) There is no registration value for mctVersion, its determined by other registration values
2.) the MCT alternative algorithms are similar enough to standard to not need separate code
