+++ src/Makefile.in
@@ -137,17 +137,17 @@ am__DEPENDENCIES_1 =
 libacvp_la_DEPENDENCIES = $(am__DEPENDENCIES_1) $(am__DEPENDENCIES_1)
 am_libacvp_la_OBJECTS = acvp.lo acvp_build_register.lo \
 	acvp_capabilities.lo acvp_operating_env.lo acvp_aes.lo \
-	acvp_des.lo acvp_hash.lo acvp_hash_mct.lo acvp_drbg.lo \
-	acvp_transport.lo acvp_util.lo parson.lo acvp_hmac.lo \
-	acvp_cmac.lo acvp_kmac.lo acvp_rsa_keygen.lo acvp_rsa_sig.lo \
-	acvp_rsa_prim.lo acvp_dsa.lo acvp_kdf135_snmp.lo \
-	acvp_kdf135_ssh.lo acvp_kdf135_srtp.lo acvp_kdf135_ikev2.lo \
-	acvp_kdf135_ikev1.lo acvp_kdf135_x942.lo acvp_kdf135_x963.lo \
-	acvp_kdf108.lo acvp_pbkdf.lo acvp_kdf_tls12.lo \
-	acvp_kdf_tls13.lo acvp_kas_ecc.lo acvp_kas_ffc.lo \
-	acvp_kas_ifc.lo acvp_kda.lo acvp_kts_ifc.lo \
-	acvp_safe_primes.lo acvp_ecdsa.lo acvp_eddsa.lo acvp_lms.lo \
-	acvp_ml_dsa.lo acvp_ml_kem.lo acvp_slh_dsa.lo
+	acvp_des.lo acvp_hash.lo acvp_drbg.lo acvp_transport.lo \
+	acvp_util.lo parson.lo acvp_hmac.lo acvp_cmac.lo acvp_kmac.lo \
+	acvp_rsa_keygen.lo acvp_rsa_sig.lo acvp_rsa_prim.lo \
+	acvp_dsa.lo acvp_kdf135_snmp.lo acvp_kdf135_ssh.lo \
+	acvp_kdf135_srtp.lo acvp_kdf135_ikev2.lo acvp_kdf135_ikev1.lo \
+	acvp_kdf135_x942.lo acvp_kdf135_x963.lo acvp_kdf108.lo \
+	acvp_pbkdf.lo acvp_kdf_tls12.lo acvp_kdf_tls13.lo \
+	acvp_kas_ecc.lo acvp_kas_ffc.lo acvp_kas_ifc.lo acvp_kda.lo \
+	acvp_kts_ifc.lo acvp_safe_primes.lo acvp_ecdsa.lo \
+	acvp_eddsa.lo acvp_lms.lo acvp_ml_dsa.lo acvp_ml_kem.lo \
+	acvp_slh_dsa.lo
 libacvp_la_OBJECTS = $(am_libacvp_la_OBJECTS)
 AM_V_lt = $(am__v_lt_@AM_V@)
 am__v_lt_ = $(am__v_lt_@AM_DEFAULT_V@)
@@ -174,10 +174,10 @@ am__depfiles_remade = ./$(DEPDIR)/acvp.Plo ./$(DEPDIR)/acvp_aes.Plo \
 	./$(DEPDIR)/acvp_des.Plo ./$(DEPDIR)/acvp_drbg.Plo \
 	./$(DEPDIR)/acvp_dsa.Plo ./$(DEPDIR)/acvp_ecdsa.Plo \
 	./$(DEPDIR)/acvp_eddsa.Plo ./$(DEPDIR)/acvp_hash.Plo \
-	./$(DEPDIR)/acvp_hash_mct.Plo ./$(DEPDIR)/acvp_hmac.Plo \
-	./$(DEPDIR)/acvp_kas_ecc.Plo ./$(DEPDIR)/acvp_kas_ffc.Plo \
-	./$(DEPDIR)/acvp_kas_ifc.Plo ./$(DEPDIR)/acvp_kda.Plo \
-	./$(DEPDIR)/acvp_kdf108.Plo ./$(DEPDIR)/acvp_kdf135_ikev1.Plo \
+	./$(DEPDIR)/acvp_hmac.Plo ./$(DEPDIR)/acvp_kas_ecc.Plo \
+	./$(DEPDIR)/acvp_kas_ffc.Plo ./$(DEPDIR)/acvp_kas_ifc.Plo \
+	./$(DEPDIR)/acvp_kda.Plo ./$(DEPDIR)/acvp_kdf108.Plo \
+	./$(DEPDIR)/acvp_kdf135_ikev1.Plo \
 	./$(DEPDIR)/acvp_kdf135_ikev2.Plo \
 	./$(DEPDIR)/acvp_kdf135_snmp.Plo \
 	./$(DEPDIR)/acvp_kdf135_srtp.Plo \
@@ -384,7 +384,6 @@ libacvp_la_SOURCES = acvp.c \
                     acvp_aes.c \
                     acvp_des.c \
                     acvp_hash.c \
-                    acvp_hash_mct.c \
                     acvp_drbg.c \
                     acvp_transport.c \
                     acvp_util.c \
@@ -515,7 +514,6 @@ distclean-compile:
 @AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/acvp_ecdsa.Plo@am__quote@ # am--include-marker
 @AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/acvp_eddsa.Plo@am__quote@ # am--include-marker
 @AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/acvp_hash.Plo@am__quote@ # am--include-marker
-@AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/acvp_hash_mct.Plo@am__quote@ # am--include-marker
 @AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/acvp_hmac.Plo@am__quote@ # am--include-marker
 @AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/acvp_kas_ecc.Plo@am__quote@ # am--include-marker
 @AMDEP_TRUE@@am__include@ @am__quote@./$(DEPDIR)/acvp_kas_ffc.Plo@am__quote@ # am--include-marker
@@ -743,7 +741,6 @@ distclean: distclean-am
 	-rm -f ./$(DEPDIR)/acvp_ecdsa.Plo
 	-rm -f ./$(DEPDIR)/acvp_eddsa.Plo
 	-rm -f ./$(DEPDIR)/acvp_hash.Plo
-	-rm -f ./$(DEPDIR)/acvp_hash_mct.Plo
 	-rm -f ./$(DEPDIR)/acvp_hmac.Plo
 	-rm -f ./$(DEPDIR)/acvp_kas_ecc.Plo
 	-rm -f ./$(DEPDIR)/acvp_kas_ffc.Plo
@@ -830,7 +827,6 @@ maintainer-clean: maintainer-clean-am
 	-rm -f ./$(DEPDIR)/acvp_ecdsa.Plo
 	-rm -f ./$(DEPDIR)/acvp_eddsa.Plo
 	-rm -f ./$(DEPDIR)/acvp_hash.Plo
-	-rm -f ./$(DEPDIR)/acvp_hash_mct.Plo
 	-rm -f ./$(DEPDIR)/acvp_hmac.Plo
 	-rm -f ./$(DEPDIR)/acvp_kas_ecc.Plo
 	-rm -f ./$(DEPDIR)/acvp_kas_ffc.Plo
